changes in the code

+++ RMS.Web/RMS.Models/Person.cs
@@ -13,7 +13,7 @@ namespace RMS.Models
         public string lastName { get; set; }
         public string Address1  { get; set; }
         public string Address2 { get; set; }
-        public int PinNo { get; set; }
+        public Int64 PinNo { get; set; }
         public int PhoneNo { get; set; }
         public Int32 MobileNo { get; set; }
         public string Email { get; set; }
